+++ src/main/java/network/crypta/node/NodeClientCore.java
@@ -25,6 +25,8 @@ import network.crypta.runtime.alerts.UserAlertManager;
 import network.crypta.runtime.alerts.UserAlertManagerClientAlertSink;
 import network.crypta.runtime.bootstrap.NodeStarter;
 import network.crypta.runtime.core.LegacyRuntimePorts;
+import network.crypta.runtime.core.NodeClientCoreSupport;
+import network.crypta.runtime.endpoints.ClientContextInitParams;
 import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.endpoints.NodeClientCoreInit;
 import network.crypta.runtime.endpoints.NodeClientPersistence;
