+++ src/test/java/network/crypta/runtime/core/NodeClientCoreSupportTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.core;
 
 import java.io.File;
 import java.io.IOException;
@@ -26,6 +26,14 @@ import network.crypta.keys.Key;
 import network.crypta.keys.SSKBlock;
 import network.crypta.keys.SSKEncodeException;
 import network.crypta.keys.SSKVerifyException;
+import network.crypta.node.ClientContextResources;
+import network.crypta.node.Node;
+import network.crypta.node.NodeClientCore;
+import network.crypta.node.NodeInitException;
+import network.crypta.node.PeerManager;
+import network.crypta.node.PeerRoutingSelectionParams;
+import network.crypta.node.PeerRoutingSelector;
+import network.crypta.node.ProgramDirectory;
 import network.crypta.runtime.alerts.DatastoreTooSmallAlert;
 import network.crypta.runtime.alerts.DiskSpaceUserAlert;
 import network.crypta.runtime.alerts.SimpleUserAlert;
