+++ src/test/java/network/crypta/runtime/endpoints/ClientEndpointsTest.java
@@ -4,9 +4,9 @@ import network.crypta.client.async.ClientContext;
 import network.crypta.config.Config;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
-import network.crypta.node.NodeClientCoreSupport;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.core.NodeClientCoreSupport;
 import network.crypta.runtime.endpoints.fcp.FcpEndpointHandle;
 import network.crypta.runtime.endpoints.http.HttpShellContainer;
 import network.crypta.runtime.spi.RuntimePorts;
